requirements.txt
Django==2.1.4
Pillow==5.4.1
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
uuid==1.30
